Move LSIF resolvers into enterprise.
